+++ worlds/stardew_valley/data/bundles_data/bundle_items_data.py
@@ -320,6 +320,7 @@ crab_cakes = BundleItem(Meal.crab_cakes)
 eggplant_parmesan = BundleItem(Meal.eggplant_parmesan)
 pumpkin_soup = BundleItem(Meal.pumpkin_soup)
 lucky_lunch = BundleItem(Meal.lucky_lunch)
+joja_cola = BundleItem(Trash.joja_cola)
 
 green_algae = BundleItem(WaterItem.green_algae)
 white_algae = BundleItem(WaterItem.white_algae)
@@ -353,6 +354,8 @@ rusty_spoon = BundleItem(Artifact.rusty_spoon)
 ancient_sword = BundleItem(Artifact.ancient_sword)
 ornamental_fan = BundleItem(Artifact.ornamental_fan)
 chipped_amphora = BundleItem(Artifact.chipped_amphora)
+strange_doll = BundleItem(Artifact.strange_doll)
+strange_doll_green = BundleItem(Artifact.strange_doll_green)
 
 prehistoric_scapula = BundleItem(Fossil.prehistoric_scapula)
 prehistoric_tibia = BundleItem(Fossil.prehistoric_tibia)
@@ -418,3 +421,6 @@ prize_ticket = BundleItem(Currency.prize_ticket)
 mystery_box = BundleItem(Consumable.mystery_box)
 gold_mystery_box = BundleItem(Consumable.gold_mystery_box, source=BundleItem.Sources.masteries)
 calico_egg = BundleItem(Currency.calico_egg)
+
+# apple_slices = BundleItem(ArtisanGood.specific_dried_fruit(Fruit.apple))
+# infinity_crown = BundleItem(Hat.infinity_crown)
